Move mapping utils into their own folder
