+++ src/ConditionalTokensMapping.template.ts
@@ -2,8 +2,8 @@ import { BigInt, BigDecimal, log } from '@graphprotocol/graph-ts'
 
 import { ConditionPreparation, ConditionResolution } from '../generated/ConditionalTokens/ConditionalTokens'
 import { Condition, Question, FixedProductMarketMaker, Category } from '../generated/schema'
-import { zero } from './constants';
-import { requireGlobal } from './global-utils';
+import { zero } from './utils/constants';
+import { requireGlobal } from './utils/global';
 
 export function handleConditionPreparation(event: ConditionPreparation): void {
   let condition = new Condition(event.params.conditionId.toHexString());
