+++ subgraph.template.yaml
@@ -163,4 +163,4 @@ templates:
       abis:
         - name: ERC20Detailed
           file: ./abis/ERC20Detailed.json
-      file: ./src/empty.ts
+      file: ./src/utils/empty.ts
